+++ CHANGELOG.md
@@ -38,6 +38,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Support for CURRENT_DATE text variable
 - Populate:
   - Basic support for regular list items (#480)
+- Position:
+  - Experimental support for gerber position files (#500)
 - Help for the error levels
 - Warnings:
   - Explain about wrong dir/output separation (#493)